Split up the file
