+++ packages/app/src/cli/models/extensions/specifications/ui_extension.ts
@@ -1,19 +1,14 @@
+import {findAllImportedFiles, createTypeDefinition, findNearestTsConfigDir, parseApiVersion} from './type-generation.js'
 import {Asset, AssetIdentifier, ExtensionFeature, createExtensionSpecification} from '../specification.js'
 import {NewExtensionPointSchemaType, NewExtensionPointsSchema, BaseSchema, MetafieldSchema} from '../schemas.js'
 import {loadLocalesConfig} from '../../../utilities/extensions/locales-configuration.js'
 import {getExtensionPointTargetSurface} from '../../../services/dev/extension/utilities.js'
 import {ExtensionInstance} from '../extension-instance.js'
 import {err, ok, Result} from '@shopify/cli-kit/node/result'
-import {fileExists, findPathUp, readFileSync} from '@shopify/cli-kit/node/fs'
-import {dirname, joinPath, relativizePath, resolvePath} from '@shopify/cli-kit/node/path'
+import {fileExists} from '@shopify/cli-kit/node/fs'
+import {joinPath} from '@shopify/cli-kit/node/path'
 import {outputContent, outputToken} from '@shopify/cli-kit/node/output'
 import {zod} from '@shopify/cli-kit/node/schema'
-import {AbortError} from '@shopify/cli-kit/node/error'
-import ts from 'typescript'
-import {init, parse} from 'es-module-lexer'
-import {createRequire} from 'module'
-
-const require = createRequire(import.meta.url)
 
 const dependency = '@shopify/checkout-ui-extensions'
 
@@ -326,198 +321,15 @@ function isRemoteDomExtension(
   config: ExtensionInstance['configuration'],
 ): config is ExtensionInstance<{api_version: string}>['configuration'] {
   const apiVersion = config.api_version
-  const [year, month] = apiVersion?.split('-').map((part: string) => parseInt(part, 10)) ?? []
-  if (!year || !month) {
+  if (!apiVersion) {
     return false
   }
-
+  const {year, month} = parseApiVersion(apiVersion) ?? {year: 0, month: 0}
   return year > 2025 || (year === 2025 && month >= 10)
 }
 
 export function getShouldRenderTarget(target: string) {
   return target.replace(/\.render$/, '.should-render')
-}
-
-function convertApiVersionToSemver(apiVersion: string): string {
-  const [year, month] = apiVersion.split('-')
-  if (!year || !month) {
-    throw new AbortError('Invalid API version format. Expected format: YYYY-MM')
-  }
-  return `${year}.${month}.0`
-}
-
-function loadTsConfig(startPath: string): {compilerOptions: ts.CompilerOptions; configPath: string | undefined} {
-  const configPath = ts.findConfigFile(startPath, ts.sys.fileExists.bind(ts.sys), 'tsconfig.json')
-  if (!configPath) {
-    return {compilerOptions: {}, configPath: undefined}
-  }
-
-  const configFile = ts.readConfigFile(configPath, ts.sys.readFile.bind(ts.sys))
-  if (configFile.error) {
-    return {compilerOptions: {}, configPath}
-  }
-
-  const parsedConfig = ts.parseJsonConfigFileContent(configFile.config, ts.sys, dirname(configPath))
-
-  return {compilerOptions: parsedConfig.options, configPath}
-}
-
-async function fallbackResolve(importPath: string, baseDir: string): Promise<string | null> {
-  // Only handle relative imports in fallback
-  if (!importPath.startsWith('./') && !importPath.startsWith('../')) {
-    return null
-  }
-
-  const resolvedPath = resolvePath(baseDir, importPath)
-  const extensions = ['', '.js', '.jsx', '.ts', '.tsx']
-
-  // Try different extensions
-  for (const ext of extensions) {
-    const pathWithExt = resolvedPath + ext
-    // eslint-disable-next-line no-await-in-loop
-    if ((await fileExists(pathWithExt)) && !pathWithExt.includes('node_modules')) {
-      return pathWithExt
-    }
-  }
-
-  // Try as directory with index files
-  for (const ext of ['.js', '.jsx', '.ts', '.tsx']) {
-    const indexPath = joinPath(resolvedPath, `index${ext}`)
-    // eslint-disable-next-line no-await-in-loop
-    if ((await fileExists(indexPath)) && !indexPath.includes('node_modules')) {
-      return indexPath
-    }
-  }
-
-  return null
-}
-
-async function parseAndResolveImports(filePath: string): Promise<string[]> {
-  try {
-    await init
-
-    const content = readFileSync(filePath).toString()
-    const resolvedPaths: string[] = []
-
-    // Load TypeScript configuration once
-    const {compilerOptions} = loadTsConfig(filePath)
-
-    const [imports] = parse(content)
-
-    const processedImports = new Set<string>()
-
-    for (const pattern of imports) {
-      const importPath = pattern.n
-
-      // Skip if already processed
-      if (!importPath || processedImports.has(importPath)) {
-        continue
-      }
-
-      processedImports.add(importPath)
-
-      // Use TypeScript's module resolution to resolve potential "paths" configurations
-      const resolvedModule = ts.resolveModuleName(importPath, filePath, compilerOptions, ts.sys)
-      if (resolvedModule.resolvedModule?.resolvedFileName) {
-        const resolvedPath = resolvedModule.resolvedModule.resolvedFileName
-
-        if (!resolvedPath.includes('node_modules')) {
-          resolvedPaths.push(resolvedPath)
-        }
-      } else {
-        // Fallback to manual resolution for edge cases
-        // eslint-disable-next-line no-await-in-loop
-        const fallbackPath = await fallbackResolve(importPath, dirname(filePath))
-        if (fallbackPath) {
-          resolvedPaths.push(fallbackPath)
-        }
-      }
-    }
-
-    return resolvedPaths
-  } catch (error) {
-    // Re-throw AbortError as-is, wrap other errors
-    if (error instanceof AbortError) {
-      throw error
-    }
-    return []
-  }
-}
-
-async function findAllImportedFiles(filePath: string, visited = new Set<string>()): Promise<string[]> {
-  if (visited.has(filePath)) {
-    return []
-  }
-
-  visited.add(filePath)
-  const resolvedPaths = await parseAndResolveImports(filePath)
-
-  const allFiles = [...resolvedPaths]
-
-  // Recursively find imports from the resolved files
-  for (const resolvedPath of resolvedPaths) {
-    // eslint-disable-next-line no-await-in-loop
-    const nestedImports = await findAllImportedFiles(resolvedPath, visited)
-    allFiles.push(...nestedImports)
-  }
-
-  return [...new Set(allFiles)]
-}
-
-function createTypeDefinition(
-  fullPath: string,
-  typeFilePath: string,
-  targets: string[],
-  apiVersion: string,
-): string | null {
-  try {
-    // Validate that all targets can be resolved
-    for (const target of targets) {
-      try {
-        require.resolve(`@shopify/ui-extensions/${target}`, {paths: [fullPath, typeFilePath]})
-      } catch (_) {
-        // Throw specific error for the target that failed, matching the original getSharedTypeDefinition behavior
-        throw new AbortError(
-          `Type reference for ${target} could not be found. You might be using the wrong @shopify/ui-extensions version.`,
-          `Fix the error by ensuring you have the correct version of @shopify/ui-extensions, for example ${convertApiVersionToSemver(
-            apiVersion,
-          )}, in your dependencies.`,
-        )
-      }
-    }
-
-    const relativePath = relativizePath(fullPath, dirname(typeFilePath))
-
-    if (targets.length === 1) {
-      const target = targets[0] ?? ''
-      return `//@ts-ignore\ndeclare module './${relativePath}' {\n  const shopify: import('@shopify/ui-extensions/${target}').Api;\n  const globalThis: { shopify: typeof shopify };\n}\n`
-    } else if (targets.length > 1) {
-      const unionType = targets.map((target) => `import('@shopify/ui-extensions/${target}').Api`).join(' | ')
-      return `//@ts-ignore\ndeclare module './${relativePath}' {\n  const shopify: ${unionType};\n  const globalThis: { shopify: typeof shopify };\n}\n`
-    }
-
-    return null
-  } catch (error) {
-    // Re-throw AbortError as-is, wrap other errors
-    if (error instanceof AbortError) {
-      throw error
-    }
-    throw new AbortError(
-      `Type reference could not be found. You might be using the wrong @shopify/ui-extensions version.`,
-      `Fix the error by ensuring you have the correct version of @shopify/ui-extensions, for example ${convertApiVersionToSemver(
-        apiVersion,
-      )}, in your dependencies.`,
-    )
-  }
-}
-
-async function findNearestTsConfigDir(fromFile: string, extensionDirectory: string): Promise<string | undefined> {
-  const fromDirectory = dirname(fromFile)
-  const tsconfigPath = await findPathUp('tsconfig.json', {cwd: fromDirectory, type: 'file', stopAt: extensionDirectory})
-
-  if (tsconfigPath) {
-    return dirname(tsconfigPath)
-  }
 }
 
 export default uiExtensionSpec
